[LamHNT] change subjectRating and teacherRating in StudentInClass to Double

+++ src/main/java/cnpr/lcss/dao/StudentInClass.java
@@ -22,9 +22,9 @@ public class StudentInClass implements Serializable {
     @Column(name = "student_class_id")
     private Integer studentInClassId;
     @Column(name = "teacher_rating")
-    private int teacherRating;
+    private double teacherRating;
     @Column(name = "subject_rating")
-    private int subjectRating;
+    private double subjectRating;
     @Column(name = "feedback")
     private String feedback;
 
@@ -40,7 +40,7 @@ public class StudentInClass implements Serializable {
     private List<Attendance> attendanceList;
 
     //<editor-fold desc="Modify Constructor">
-    public StudentInClass(Integer studentInClassId, int teacherRating, int subjectRating, String feedback) {
+    public StudentInClass(Integer studentInClassId, double teacherRating, double subjectRating, String feedback) {
         this.studentInClassId = studentInClassId;
         this.teacherRating = teacherRating;
         this.subjectRating = subjectRating;
